Avoid another cyclic import
